creditmanagement/requirements.txt
amqp==5.1.1
aniso8601==9.0.1
asgiref==3.5.2
async-timeout==4.0.2
autopep8==2.0.1
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
datetime-truncate==1.1.1
dj-database-url==1.2.0
Django==4.1.4
django-admin-charts==1.2.4
django-admin-interface==0.24.2
django-bower==5.2.0
django-colorfield==0.8.0
django-cors-headers==3.13.0
django-cloudinary-storage==0.3.0
django-dotenv==1.4.2
django-easy-pjax==1.3.0
django-elasticsearch-dsl==7.2.2
django-environ==0.9.0
django-filter==22.1
django-memoize==2.3.1
django-multiselectfield==0.1.12
django-nvd3==0.9.7
django-upgrade==1.12.0
django-utils-six==2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.2.1
elasticsearch==7.17.8
elasticsearch-dsl==7.4.0
email-validator==1.3.0
eventlet==0.33.3
graphene==3.2.1
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.2
Pillow==9.3.0
portalocker==2.7.0
promise==2.3
prompt-toolkit==3.0.38
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycodestyle==2.10.0
PyJWT==2.6.0
python-dateutil==2.8.2
python-decouple==3.7
python-nvd3==0.14.2
python-slugify==1.1.4
pytz==2022.6
redis==4.5.1
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
tokenize-rt==5.0.0
tomli==2.0.1
tzdata==2022.7
Unidecode==1.3.6
urllib3==1.26.14
vine==5.0.0
# whitenoise==5.1.0
waitress==2.1.2
wcwidth==0.2.6
xhtml2pdf==0.2.1



# amqp==5.1.1
# aniso8601==9.0.1
# asgiref==3.5.2
# async-timeout==4.0.2
# autopep8==2.0.1
# backports.zoneinfo==0.2.1
# billiard==3.6.4.0
# certifi==2022.12.7
# channels==3.0.4
# channels-redis==3.3.1
# click==8.1.3
# click-didyoumean==0.3.0
# click-plugins==1.1.1
# click-repl==0.2.0
# colorama==0.4.6
# datetime-truncate==1.1.1
# dj-database-url==1.2.0
# Django==4.1.4
# django-admin-charts==1.2.4
# django-admin-interface==0.24.2
# django-bower==5.2.0
# django-colorfield==0.8.0
# django-cors-headers==3.13.0
# django-cloudinary-storage==0.3.0
# django-dotenv==1.4.2
# django-easy-pjax==1.3.0
# django-elasticsearch-dsl==7.2.2
# django-environ==0.9.0
# django-filter==22.1
# django-memoize==2.3.1
# django-multiselectfield==0.1.12
# django-nvd3==0.9.7
# django-upgrade==1.12.0
# django-utils-six==2.0
# djangorestframework==3.14.0
# djangorestframework-simplejwt==5.2.2
# dnspython==2.2.1
# elasticsearch==7.17.8
# elasticsearch-dsl==7.4.0
# email-validator==1.3.0
# eventlet==0.33.3
# graphene==3.2.1
# graphene-django==3.0.0
# graphql-core==3.2.3
# graphql-relay==3.2.0
# greenlet==2.0.2
# gunicorn==20.1.0
# idna==3.4
# Jinja2==3.1.2
# kombu==5.2.4
# MarkupSafe==2.1.2
# Pillow==9.3.0
# portalocker==2.7.0
# promise==2.3
# prompt-toolkit==3.0.38
# psycopg2==2.9.5
# psycopg2-binary==2.9.5
# pycodestyle==2.10.0
# PyJWT==2.6.0
# python-dateutil==2.8.2
# python-decouple==3.7
# python-nvd3==0.14.2
# python-slugify==1.1.4
# pytz==2022.6
# redis==4.5.5
# six==1.16.0
# sqlparse==0.4.3
# text-unidecode==1.3
# tokenize-rt==5.0.0
# tomli==2.0.1
# tzdata==2022.7
# Unidecode==1.3.6
# urllib3==1.26.14
# vine==5.0.0
# # whitenoise==5.1.0
# waitress==2.1.2
# wcwidth==0.2.6
# xhtml2pdf==0.2.11